+++ react/components/LayoutModeSwitcher.js
@@ -2,14 +2,8 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { Button } from 'vtex.styleguide'
 import { IconGrid, IconInlineGrid, IconSingleGrid } from 'vtex.store-icons'
-import { prop } from 'ramda'
 
 import searchResult from '../searchResult.css'
-import {
-  useSearchPage,
-  useSearchPageState,
-  useSearchPageStateDispatch,
-} from 'vtex.search-page-context/SearchPageContext'
 
 export const LAYOUT_MODE = [
   {
@@ -73,29 +67,4 @@ LayoutModeSwitcher.propTypes = {
   onChange: PropTypes.func.isRequired,
 }
 
-const withSearchPageContextProps = Component => props => {
-  const { mobileLayout: layoutOptions } = useSearchPage()
-  const { mobileLayout } = useSearchPageState()
-  const dispatch = useSearchPageStateDispatch()
-
-  const onChange = e => {
-    e.preventDefault()
-    const newLayoutIndex =
-      prop('mode1', layoutOptions) === mobileLayout ? 'mode2' : 'mode1'
-    const newLayout = prop(newLayoutIndex, layoutOptions)
-    dispatch({ type: 'SWITCH_LAYOUT', args: { mobileLayout: newLayout } })
-  }
-
-  if (layoutOptions && Object.keys(layoutOptions).length < 2) {
-    return null
-  }
-
-  return (
-    <Component
-      activeMode={props.activeMode || mobileLayout}
-      onChange={props.onChange || onChange}
-    />
-  )
-}
-
-export default withSearchPageContextProps(LayoutModeSwitcher)
+export default LayoutModeSwitcher
