+++ react/searchResult.css
@@ -340,3 +340,6 @@
 
 .container--layout {
 }
+
+.notFound--layout {
+}
